requirements.txt
gymnasium
numpy
torch
lovely_tensors
pygame
faiss-cpu
sentence_transformers
timm 
torchvision
stable-baselines3